# [#21] Unpack fields